a few last minute changes brought to my attention from Tom and Joe.
A thing that needs special attention is run_time_settings.py, the tooltip for 'volume_path' on line 30.

This should be good for now, there are some future concerns which we might wish to address in the future such as page adjusting to varying aspect ratios and people using differing degrees of zoom in their browser. However we are assuing users emply 'standard' aspect ratios and zoom setting.