there are currently a lot of errors with the background code, doing my best with what I have.
preview works except for displaying the additional figures if wrmXpress finishes the image analysis, wrmXpress currently fails in the background, this needs to be fixed.

run works for (motility and segmentation) analyzing the individual wells as they are being processed however the wrmXpress still fails in the background and there is no finished stitched image so this doesnt diplay when finished.

these pages are currently working for motility and segmentation and are built out the best they can given the limitations in the background code in wrmXpress. will move onto cellprofiler modules.